requirements.txt
asgiref==3.3.4
Django==3.2
django-logic==0.1.2
djangorestframework==3.12.4
graphviz==0.16
pytz==2021.1
sqlparse==0.4.1
